requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
assemblyai==0.34.0
backports.zoneinfo==0.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==4.2.16
exceptiongroup==1.2.2
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
joblib==1.4.2
nltk==3.9.1
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.1
textblob==0.18.0.post0
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
